small parser improvements
