+++ packages/language-server/src/language-service/parser/index.ts
@@ -1,22 +1,6 @@
-import { Referenceables } from "./../model/referenceables"
-import {
-	CUSTOM_TAGS,
-	CUSTOM_TAGS_BY_PROPERTY_NAME,
-	CustomTag
-} from "./../model/custom-tags"
+import { CUSTOM_TAGS } from "./../model/custom-tags"
 
-import {
-	ArrayASTNode,
-	ASTNode,
-	BooleanASTNode,
-	ErrorCode,
-	JSONDocument,
-	NullASTNode,
-	NumberASTNode,
-	ObjectASTNode,
-	PropertyASTNode,
-	StringASTNode
-} from "./json"
+import { YAMLDocument, ErrorCode, Problem } from "./json"
 
 import * as nls from "vscode-nls"
 const localize = nls.loadMessageBundle()
@@ -24,280 +8,9 @@ const localize = nls.loadMessageBundle()
 import { Schema, Type } from "js-yaml"
 import * as Yaml from "yaml-ast-parser"
 
-import { GlobalsConfig } from "../model/globals"
-import { collectGlobals, getDefaultGlobalsConfig } from "./globals"
-import { parseYamlBoolean } from "./scalar-type"
-import { ResourcesDefinitions } from "../model/resources"
-import { DocumentType } from "../model/document"
 import { getDocumentType } from "../utils/document"
-import { collectReferenceables } from "./referenceables"
-import { collectReferences } from "./references"
-import { Reference } from "../model/references"
 
-export interface Problem {
-	message: string
-	location: {
-		start: number
-		end: number
-	}
-	code: ErrorCode
-}
-
-interface YamlScalar extends Yaml.YAMLScalar {
-	customTag?: CustomTag
-}
-
-interface YamlSequence extends Yaml.YAMLSequence {
-	customTag?: CustomTag
-}
-
-export class YAMLDocument extends JSONDocument {
-	root: ASTNode
-	errors: Problem[]
-	warnings: Problem[]
-	globalsConfig: GlobalsConfig
-	resources: ResourcesDefinitions
-	referenceables: Referenceables
-	references: Reference[]
-	parameters: string[]
-	documentType: DocumentType
-
-	constructor(documentType: DocumentType) {
-		super(null, [])
-		this.documentType = documentType
-		this.globalsConfig = getDefaultGlobalsConfig()
-		this.root = null
-		this.errors = []
-		this.warnings = []
-	}
-
-	getSchemas(schema, doc, node: ASTNode) {
-		const matchingSchemas = []
-		doc.validate(schema, matchingSchemas, node.start)
-		return matchingSchemas
-	}
-
-	getNodeFromOffset(offset: number): ASTNode {
-		return this.getNodeFromOffsetEndInclusive(offset)
-	}
-}
-
-function recursivelyBuildAst(parent: ASTNode, node: Yaml.YAMLNode): ASTNode {
-	if (!node) {
-		return
-	}
-
-	switch (node.kind) {
-		case Yaml.Kind.MAP: {
-			const instance = node as Yaml.YamlMap
-
-			const result = new ObjectASTNode(
-				parent,
-				null,
-				node.startPosition,
-				node.endPosition
-			)
-
-			for (const mapping of instance.mappings) {
-				result.addProperty(recursivelyBuildAst(
-					result,
-					mapping
-				) as PropertyASTNode)
-			}
-
-			return result
-		}
-		case Yaml.Kind.MAPPING: {
-			const instance = node as Yaml.YAMLMapping
-			const key = instance.key
-			const customTag = CUSTOM_TAGS_BY_PROPERTY_NAME[key.value]
-
-			// Technically, this is an arbitrary node in YAML
-			// I doubt we would get a better string representation by parsing it
-			const keyNode = new StringASTNode(
-				null,
-				null,
-				true,
-				key.startPosition,
-				key.endPosition
-			)
-			keyNode.value = key.value
-
-			const result = new PropertyASTNode(parent, keyNode, customTag)
-			result.end = instance.endPosition
-
-			const valueNode = instance.value
-				? recursivelyBuildAst(result, instance.value)
-				: new NullASTNode(
-						parent,
-						key.value,
-						instance.endPosition,
-						instance.endPosition
-				  )
-			valueNode.location = key.value
-
-			result.setValue(valueNode)
-
-			return result
-		}
-		case Yaml.Kind.SEQ: {
-			const instance = node as YamlSequence
-
-			const result = new ArrayASTNode(
-				parent,
-				null,
-				instance.startPosition,
-				instance.endPosition,
-				instance.customTag
-			)
-
-			let count = 0
-			for (const item of instance.items) {
-				if (item === null && count === instance.items.length - 1) {
-					break
-				}
-
-				// Be aware of https://github.com/nodeca/js-yaml/issues/321
-				// Cannot simply work around it here because we need to know if we are in Flow or Block
-				const itemNode =
-					item === null
-						? new NullASTNode(
-								parent,
-								null,
-								instance.endPosition,
-								instance.endPosition
-						  )
-						: recursivelyBuildAst(result, item)
-
-				itemNode.location = count++
-				result.addItem(itemNode)
-			}
-
-			return result
-		}
-		case Yaml.Kind.SCALAR: {
-			const instance = node as YamlScalar
-			const type = Yaml.determineScalarType(instance)
-
-			// The name is set either by the sequence or the mapping case.
-			const name = null
-			const value = instance.value
-
-			// This is a patch for redirecting values with these strings to be boolean nodes because its not supported in the parser.
-			const possibleBooleanValues = [
-				"y",
-				"Y",
-				"yes",
-				"Yes",
-				"YES",
-				"n",
-				"N",
-				"no",
-				"No",
-				"NO",
-				"on",
-				"On",
-				"ON",
-				"off",
-				"Off",
-				"OFF"
-			]
-			if (
-				instance.plainScalar &&
-				possibleBooleanValues.indexOf(value.toString()) !== -1
-			) {
-				return new BooleanASTNode(
-					parent,
-					name,
-					parseYamlBoolean(value),
-					node.startPosition,
-					node.endPosition
-				)
-			}
-
-			switch (type) {
-				case Yaml.ScalarType.null: {
-					return new StringASTNode(
-						parent,
-						name,
-						false,
-						instance.startPosition,
-						instance.endPosition
-					)
-				}
-				case Yaml.ScalarType.bool: {
-					return new BooleanASTNode(
-						parent,
-						name,
-						Yaml.parseYamlBoolean(value),
-						node.startPosition,
-						node.endPosition
-					)
-				}
-				case Yaml.ScalarType.int: {
-					const result = new NumberASTNode(
-						parent,
-						name,
-						node.startPosition,
-						node.endPosition
-					)
-					result.value = Yaml.parseYamlInteger(value)
-					result.isInteger = true
-					return result
-				}
-				case Yaml.ScalarType.float: {
-					const result = new NumberASTNode(
-						parent,
-						name,
-						node.startPosition,
-						node.endPosition
-					)
-					result.value = Yaml.parseYamlFloat(value)
-					result.isInteger = false
-					return result
-				}
-				case Yaml.ScalarType.string: {
-					const result = new StringASTNode(
-						parent,
-						name,
-						false,
-						node.startPosition,
-						node.endPosition,
-						instance.customTag
-					)
-					result.value = node.value
-					return result
-				}
-			}
-
-			break
-		}
-		case Yaml.Kind.ANCHOR_REF: {
-			const instance = (node as Yaml.YAMLAnchorReference).value
-
-			return (
-				recursivelyBuildAst(parent, instance) ||
-				new NullASTNode(
-					parent,
-					null,
-					node.startPosition,
-					node.endPosition
-				)
-			)
-		}
-		case Yaml.Kind.INCLUDE_REF: {
-			const result = new StringASTNode(
-				parent,
-				null,
-				false,
-				node.startPosition,
-				node.endPosition
-			)
-			result.value = node.value
-			return result
-		}
-	}
-}
+export { YAMLDocument, Problem }
 
 function convertError(e: Yaml.YAMLException): Problem {
 	return {
@@ -311,13 +24,7 @@ function convertError(e: Yaml.YAMLException): Problem {
 }
 
 function createJSONDocument(yamlDoc: Yaml.YAMLNode | void, text: string) {
-	const doc = new YAMLDocument(getDocumentType(text))
-
-	if (yamlDoc) {
-		doc.root = recursivelyBuildAst(null, yamlDoc)
-		doc.referenceables = collectReferenceables(doc.documentType, doc.root)
-		doc.references = collectReferences(doc.root)
-	}
+	const doc = new YAMLDocument(getDocumentType(text), yamlDoc)
 
 	if (!yamlDoc || !doc.root) {
 		// TODO: When this is true, consider not pushing the other errors.
@@ -337,8 +44,6 @@ function createJSONDocument(yamlDoc: Yaml.YAMLNode | void, text: string) {
 
 		return doc
 	}
-
-	doc.globalsConfig = collectGlobals(doc.root)
 
 	const duplicateKeyReason = "duplicate key"
