+++ packages/language-server/src/language-service/parser/json/boolean-ast-node.ts
@@ -1,3 +1,4 @@
+import { YAMLDocument } from "./../index"
 import { ASTNode } from "./ast-node"
 import * as Json from "jsonc-parser"
 
@@ -5,13 +6,14 @@ export class BooleanASTNode extends ASTNode {
 	private value: boolean | string
 
 	constructor(
+		document: YAMLDocument,
 		parent: ASTNode,
 		name: Json.Segment,
 		value: boolean | string,
 		start: number,
 		end?: number
 	) {
-		super(parent, "boolean", name, start, end)
+		super(document, parent, "boolean", name, start, end)
 		this.value = value
 	}
