+++ packages/language-server/src/language-service/parser/json/null-ast-node.ts
@@ -1,14 +1,16 @@
+import { YAMLDocument } from "./../index"
 import { ASTNode } from "./ast-node"
 import * as Json from "jsonc-parser"
 
 export class NullASTNode extends ASTNode {
 	constructor(
+		document: YAMLDocument,
 		parent: ASTNode,
 		name: Json.Segment,
 		start: number,
 		end?: number
 	) {
-		super(parent, "null", name, start, end)
+		super(document, parent, "null", name, start, end)
 	}
 
 	getValue(): null {
