+++ packages/language-server/src/language-service/parser/util.ts
@@ -21,6 +21,6 @@ export const getPropertyNodeValue = (
 	}
 
 	return propertyNode.getChildNodes().find(node => {
-		return node.type === "object" && node.location === location
+		return node instanceof ObjectASTNode && node.location === location
 	}) as ObjectASTNode | void
 }
