+++ packages/language-server/src/language-service/services/jsonSchema/__tests__/mutation.test.ts
@@ -1,11 +1,12 @@
+import { UNKNOWN } from "./../../../model/document"
 import { ResolvedSchema } from ".."
 import { FUNCTION } from "../../../model/globals"
 import { YAMLDocument } from "../../../parser"
-import { getDefaultGlobalsConfig } from "../../../parser/globals"
 import { applyDocumentMutations } from "../mutation"
+import { getDefaultGlobalsConfig } from "../../../parser/json/globals"
 
 test("should do nothing if schema was not resolved", () => {
-	const doc = new YAMLDocument([])
+	const doc = new YAMLDocument(UNKNOWN, undefined)
 	doc.globalsConfig = getDefaultGlobalsConfig()
 
 	expect(applyDocumentMutations(undefined, doc)).toBeUndefined()
@@ -15,7 +16,7 @@ test("should return original schema if globals are not defined", () => {
 	const emptySchema = {}
 	const emptyGlobals = getDefaultGlobalsConfig()
 	const resolvedSchema = new ResolvedSchema(emptySchema)
-	const doc = new YAMLDocument([])
+	const doc = new YAMLDocument(UNKNOWN, undefined)
 	doc.globalsConfig = emptyGlobals
 
 	expect(applyDocumentMutations(resolvedSchema, doc)).toBe(resolvedSchema)
@@ -27,7 +28,7 @@ test("should return original schema if definitions are defined", () => {
 	}
 	const emptyGlobals = getDefaultGlobalsConfig()
 	const resolvedSchema = new ResolvedSchema(schema)
-	const doc = new YAMLDocument([])
+	const doc = new YAMLDocument(UNKNOWN, undefined)
 	doc.globalsConfig = emptyGlobals
 
 	expect(applyDocumentMutations(resolvedSchema, doc)).toBe(resolvedSchema)
@@ -48,7 +49,7 @@ test("should return schema with mutations for existing resource", () => {
 	const globals = getDefaultGlobalsConfig()
 	globals[FUNCTION].properties = ["Runtime"]
 	const resolvedSchema = new ResolvedSchema(schema)
-	const doc = new YAMLDocument([])
+	const doc = new YAMLDocument(UNKNOWN, undefined)
 	doc.globalsConfig = globals
 
 	const newResolvedSchema = applyDocumentMutations(
